+++ packages/babylon.quarks/src/shaders/chunks/soft_fragment.glsl.ts
@@ -5,7 +5,7 @@ export default /* glsl */ `
     p2 = 0.5 * p2 + 0.5;
 
     float readDepth = texture2D(depthTexture, p2.xy).r;
-    float viewDepth = linearize_depth(readDepth);
+    float viewDepth = decode_depth(readDepth);
     float softParticlesFade = saturate(SOFT_INV_FADE_DISTANCE * ((viewDepth - SOFT_NEAR_FADE) - linearDepth));
 
     gl_FragColor *= softParticlesFade;